first test of item selection
(I need a controller lel)

+++ Wiwa/src/Wiwa/Ui/UiManager.h
@@ -75,7 +75,7 @@ namespace Wiwa
 
 		Image* texture;
 
-		size_t idGuiSelected = -1;
+		int idGuiSelected = -1;
 
 	};
 }
